backend/requirements_lean.txt
# ARIA WAR MACHINE - Core Dependencies Only
# No fluff, no docs, no UI

# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# MT5
MetaTrader5==5.0.45

# Redis
redis==5.0.1
aioredis==2.0.1

# ML/AI
numpy==1.24.3
pandas==2.1.3
scikit-learn==1.3.2
xgboost==2.0.2
onnxruntime==1.16.3

# Async
asyncio==3.4.3
aiofiles==23.2.1

# Utils
pydantic==2.5.0
python-multipart==0.0.6
